Move module`s routes into his modules
