+++ src/app/components/courses-for-groups/courses-for-groups.module.ts
@@ -1,7 +1,6 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { CoursesForGroupsComponent } from './courses-for-groups.component';
-import {HttpClientModule} from '@angular/common/http';
 import {FormsModule, ReactiveFormsModule} from '@angular/forms';
 import { StudiedCoursesComponent } from './studied-courses/studied-courses.component';
 import { AddedCoursesComponent } from './added-courses/added-courses.component';
@@ -11,8 +10,11 @@ import { CopyCoursesDialogComponent } from './copy-courses-dialog/copy-courses-d
 import {NgbModule} from '@ng-bootstrap/ng-bootstrap';
 import {SharedModule} from '../shared/shared.module';
 import {PipeModule} from '../../pipes/pipe.module';
-import {RouterModule} from '@angular/router';
-import {coursesForGroupsRoutes} from '../../routes/routes';
+import {RouterModule, Routes} from '@angular/router';
+
+const coursesForGroupsRoutes: Routes = [
+  {path: '', component: CoursesForGroupsComponent}
+];
 
 @NgModule({
   imports: [
