+++ src/app/components/diploma-supplement/diploma-supplement.module.ts
@@ -2,9 +2,12 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import {FormsModule} from '@angular/forms';
 import {SharedModule} from '../shared/shared.module';
-import {RouterModule} from '@angular/router';
-import {diplomaSupplementRoutes} from '../../routes/routes';
+import {RouterModule, Routes} from '@angular/router';
 import {DiplomaSupplementComponent} from './diploma-supplement.component';
+
+const diplomaSupplementRoutes: Routes = [
+  {path: '', component: DiplomaSupplementComponent}
+];
 
 @NgModule({
   imports: [
