+++ src/app/components/login/authentication.module.ts
@@ -2,9 +2,12 @@ import {NgModule} from '@angular/core';
 import {LoginComponent} from './login.component';
 import {FormsModule} from '@angular/forms';
 import {HttpClientModule} from '@angular/common/http';
-import {RouterModule} from '@angular/router';
-import {authRoutes} from '../../routes/routes';
+import {RouterModule, Routes} from '@angular/router';
 import {CommonModule} from '@angular/common';
+
+export const authRoutes: Routes = [
+  {path: '', component: LoginComponent}
+];
 
 @NgModule({
   declarations: [
