+++ src/app/components/speciality/speciality.module.ts
@@ -2,9 +2,12 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import {FormsModule} from '@angular/forms';
 import {SpecialityComponent} from './speciality.component';
-import {RouterModule} from '@angular/router';
-import {specialityRoutes} from '../../routes/routes';
+import {RouterModule, Routes} from '@angular/router';
 import {SharedModule} from '../shared/shared.module';
+
+export const specialityRoutes: Routes = [
+  {path: '', component: SpecialityComponent}
+];
 
 @NgModule({
   imports: [
